requirements.txt
openbb
psycopg2-binary
openai
python-dotenv
